requirements.txt
appdirs==1.4.3
cycler==0.10.0
funcsigs==1.0.2
functools32==3.2.3.post2
matplotlib==2.0.0
mock==2.0.0
numpy==1.12.1
olefile==0.44
packaging==16.8
pandas==0.19.2
pbr==3.0.0
Pillow==4.1.0
protobuf==3.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
scipy==0.19.0
six==1.10.0
subprocess32==3.2.7
tensorflow-gpu==1.1.0
tensorvision==0.1.dev1
Werkzeug==0.12.1
